feat(signup): linked password strength to input wrning and separate into modules the signup validation
