+++ src/components/PasswordStrengthMeter/styles.module.scss
@@ -1,0 +1,12 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 16;
+  color: #161616;
+}
+
+.check {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
