+++ src/utils/SignUpValidations/index.ts
@@ -1,0 +1,12 @@
+import { validateEmail } from "./email";
+import { validateFirstName } from "./firstName";
+import { validateLastName } from "./lastName";
+import { validatePassword, isPasswordStrong } from "./password";
+
+export {
+  validateEmail,
+  validateFirstName,
+  validateLastName,
+  validatePassword,
+  isPasswordStrong,
+};
