+++ src/utils/SignUpValidations/lastName.ts
@@ -1,0 +1,15 @@
+import { FormErrors } from "./types";
+
+export const validateLastName = (
+  lastName: string,
+  setFormErrors: React.Dispatch<React.SetStateAction<FormErrors>>
+) => {
+  let errorMessage = "";
+  if (!lastName) {
+    errorMessage = "Last name is required";
+  }
+  setFormErrors((prevFormErrors) => ({
+    ...prevFormErrors,
+    lastName: errorMessage,
+  }));
+};
