+++ src/utils/SignUpValidations/types.ts
@@ -1,0 +1,13 @@
+export interface UserData {
+  [key: string]: string;
+}
+
+export interface FormErrors {
+  [key: string]: string;
+}
+
+export interface PasswordRequirements {
+  hasMinLength: boolean;
+  hasLetters: boolean;
+  hasNumbers: boolean;
+}
